requirements.txt
Django==4.2.15
djangorestframework==3.15.2
django-filter
